fix: count(*)::bigint to match the hub's bigint columns (MC-2368 review)
Hub columns are bigint; the fortress cast count(*)::int throws past ~2.1B rows
and would swallow the whole stats push. scalarInt's Number() already parses the
bigint string.

+++ src/query/collection-stats.ts
@@ -21,10 +21,10 @@ function scalarInt(res: unknown): number {
 
 export async function computeCollectionStats(db: HxDb): Promise<CollectionStats> {
   const sessions = scalarInt(
-    await db.execute(sql`SELECT count(*)::int AS n FROM hx.sessions WHERE deleted_at IS NULL`),
+    await db.execute(sql`SELECT count(*)::bigint AS n FROM hx.sessions WHERE deleted_at IS NULL`),
   );
   const turns = scalarInt(
-    await db.execute(sql`SELECT count(*)::int AS n FROM hx.turns WHERE deleted_at IS NULL`),
+    await db.execute(sql`SELECT count(*)::bigint AS n FROM hx.turns WHERE deleted_at IS NULL`),
   );
 
   // to_regclass returns NULL (never errors) when hx.embeddings is absent, so the
@@ -34,7 +34,7 @@ export async function computeCollectionStats(db: HxDb): Promise<CollectionStats>
   const present = (reg[0] as { rel?: string | null } | undefined)?.rel != null;
   if (present) {
     embeddings = scalarInt(
-      await db.execute(sql`SELECT count(*)::int AS n FROM hx.embeddings WHERE deleted_at IS NULL`),
+      await db.execute(sql`SELECT count(*)::bigint AS n FROM hx.embeddings WHERE deleted_at IS NULL`),
     );
   }
